refactor input component
